moved accountinfo and bakerinfo from types.hs to types.accounts
