+++ haskell-src/Concordium/Types/Accounts.hs
@@ -9,26 +9,13 @@ import Data.Serialize
 import qualified Data.Map as Map
 import qualified Lens.Micro.Platform as Lens
 import Concordium.Common.Version
+import Concordium.Types.Accounts.Releases
 -- import Concordium.GlobalState.Account
 -- import Concordium.GlobalState.Basic.BlockState.AccountReleaseSchedule (AccountReleaseSchedule (..))
 -- import Concordium.GlobalState.Basic.BlockState.Updates (Updates)
 -- import Concordium.GlobalState.Finalization
 import Concordium.ID.Types
-    ( AccountCredential,
-      AccountEncryptionKey,
-      AccountThreshold,
-      CredentialIndex )
 import Concordium.Types
-    ( Epoch,
-      Amount,
-      AccountEncryptedAmount,
-      Nonce,
-      AccountIndex,
-      BakerAggregationVerifyKey,
-      BakerElectionVerifyKey,
-      BakerSignVerifyKey,
-      BakerId,
-      AccountReleaseSchedule )
 import Concordium.Types.HashableTo
 import qualified Concordium.Crypto.SHA256 as Hash
